Periodograms

Time varying periodograms that can be used as data input in dynamic models that uses spectral likelihoods.

Moving periodograms

Based on the periodogram in Häfner and Kirch (2017) - "Moving Fourier Analysis for Locally Stationary Processes with the Bootstrap in View". Consider adjustments in form of tapering, pre-whitening and boundary correction.

TimeVaryingPeriodograms.mvPeriodogram — Function

  mvPeriodogram(y, m, startFreq) 

Computes the moving local periodogram of the time series y using a window size m, starting at frequency startFreq.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPeriodogram(y, 10, 2)

TimeVaryingPeriodograms.mvTapPeriodogram — Function

mvTapPeriodogram(y, m, α=1, startFreq=1) 

Same as mvPeriodogram but with a Tukey taper applied to the data. Use α=1 for Hanning taper.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvTapPeriodogramDiffStartFreq(y, 10)

TimeVaryingPeriodograms.mvPrewhitePeriodogram — Function

  mvPrewhitePeriodogram(y, m, p=nothing, startFreq) 

Same as mvPeriodogram but with a Tukey taper applied to the data. Use α=1 for Hanning taper.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvTapPeriodogramDiffStartFreq(y, 10)

TimeVaryingPeriodograms.mvPrewhiteTapPeriodogram — Function

  mvPrewhiteTapPeriodogram(y, m, p, startFreq) 

Same as mvPrewhitePeriodogram but with prewhitening. p is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPreWPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.BoundCorrectedMvPeriodogram — Function

    BoundCorrectedMvPeriodogram(y, m, p, startFreq) 

Same as mvPeriodogram but with boundary correction in Rao and Yang (2021). ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = BoundCorrectedMvPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.tapBoundCorrectedMvPeriodogram — Function

tapBoundCorrectedMvPeriodogram(y, m, p, startFreq)

Same as ´BoundCorrectedMvPeriodogram´ but with tapering using a Hanning window.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = tapBoundCorrectedMvPeriodogram(y, 10, 2, 1)

Block periodograms

Based on the periodogram used in block-Whittle likelihood, see e.g. Dahlhaus (2012) - "Locally Stationary Processes" using the same adjustments as above.

TimeVaryingPeriodograms.tvPeriodogram — Function

    tvPeriodogram(y, N, S, nFreq = nothing, taper = nothing; taperArgs...)

Computes the block-periodogram (Dahlhaus, 2012) of the time series y using a window size N, and step-size S on the Fourier frequencies. NFreq can be used to evaluate other frequencies and taper can be used to apply a taper to the data, see package DSP for details.

Returns a spectrogram of the data.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia>  tvPeriodogram(y, 25, 15)

TimeVaryingPeriodograms.preWhiteTvBlockPeriodogram — Function

    preWhiteTvBlockPeriodogram(y,N,S,p=nothing)

Same as tvPeriodogram but with pre-whitening. ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = preWhiteTvBlockPeriodogram(y,25,10,p=nothing)

TimeVaryingPeriodograms.tapBoundCorrectTvBlockPeriodogram — Function

    tapBoundCorrectTvBlockPeriodogram(y,N,S,p=nothing,α=nothing,thresH=0.1^8)

Same as tvPeriodogram but with boundary correction from Rao and Yang (2021). ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = tapBoundCorrectTvBlockPeriodogram(y,25,10,p=nothing,α=nothing,thresH=0.1^8)

TimeVaryingPeriodograms.reScaleTapTvBlockPeriodogram — Function

    reScaleTapTvBlockPeriodogram(y,N,S)

Same as tvPeriodogram but with rescaling of the tapered data to match the variance of the non-scaled data. Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = reScaleTapTvBlockPeriodogram(y,25,10)

Pre-periodograms

Preperiodogram based on Neumann and Von Sachs (1997) and Everitt et al. (2013).

TimeVaryingPeriodograms.prePeriodogram — Function

prePeriodogram(y, ω)

ω is a selected vector of frequencies. Returns a (length(ω) x T-1) matrix containing the prePeriodogram.

TimeVaryingPeriodograms.EverittI — Function

EverittI(y, ωgrid, ρ, ν)

Everitt et al. version of the pre-periodogram.

y - input time series,

ωgrid - grid of frequencies to be evaluated,

ρ - Parameter that controls smoothing ove time,

ν - parameter that controlls exponential taper.

Time invariant periodograms

Time invariant periodograms that are used as parts of the above. Can be interesting for comparison.

TimeVaryingPeriodograms.whitePeriodogram — Function

whitePeriodogram(y, p)

Returns the periodogram from the prewhitening approach using AR-order p.

TimeVaryingPeriodograms.completePeriodogram — Function

   completePeriodogram(y; fr=nothing, p = nothing, ϕ = nothing, α=nothing, threshold=0.1^8)

Computing the complete periodogram on frequencies, fr (Fourier frquencies if nothing). This is needed in the boundary correction for the dynamic Whittle likellihood.

Returns the complete periodogram of the data given the model.