Utils

TimeVaryingPeriodograms.predictiveDFT — Function

  predictiveDFT(y, ω; p=nothing, ϕ=nothing)

Predictive DFT from Rao and Yang (2021), see r-package "cspec" on CRAN. Default is fitting the best AR model and compute the predictive DFT based on the predictions att frequencies ´ω´. ´p´ will fix the AR-order and ´ϕ´ gives the AR-coefficients.

Returns the discrete Fourier coefficients associated with the predictions.

Examples

julia> rand(Normal(500))
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPreWPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.completeDFTSingle — Function

   completeDFTSingle(y, fr; p = nothing, ϕ = nothing)

Computing the complete DFT on a single frequency, fr. This is needed in the boundary correction for the dynamic Whittle likellihood.

Returns the discrete Fourier transform of the data.

TimeVaryingPeriodograms.SingleDft2 — Function

SingleDft2(y, fr)

Computing the DFT on a single frequency, fr. This is needed in the boundary correction for the dynamic Whittle likellihood..

Returns the discrete Fourier transform of the data.

TimeVaryingPeriodograms.SpecDensARMA — Function

SpecDensARMA(ω, ϕ, θ, σ²)

Compute spectral density for the univariate ARMA model.

ω is a radial frequency

ϕ is a vector of AR coefficients

θ is a vector of MA coefficients

σ² is the noise variance

Examples

The spectral density for an AR(1) process with unit noise variance is

julia> SpecDensARMA(0.5, 0.9, 0, 1)
0.6909224383713601

TimeVaryingPeriodograms.wikleTV6AR — Function

wikleTV6AR(T, σ, A, μ)

Simulate spectrogram and time series from a time-varying autoregressive model of order 6.

Length of time series is T, innovation standard deviation is σ, and the AR coefficients are compute from the three-element vector A. The mean of the time series is μ.

Examples

julia> T = 1024; σ = 2; A = [1.1, 1.12, 1.1]; μ = 0;
julia> y, spdMat, coeffϕ = wikleTV6AR(T, σ, A, μ)
julia> heatmap(log.(spdMat'), c = :viridis, clim = (-3,1.5), 
    size = (1000,400), title="True Spectrogram");
julia> plot(y, label = nothing, title = "Simulated time series from TvAR(6) model", 
        xlabel = "time", ylabel = "y")

TimeVaryingPeriodograms.DahlhausAR2 — Function

DahlhausAR2(T, σ, μ)

Simulate spectrogram and time series from a time-varying autoregressive model of order 2 from Dahlhaus handbook chapter..

Length of time series is T, innovation standard deviation is σ. The mean of the time series is μ(default is 0).

Examples

julia> T = 1024; σ = 2; μ = 0;
julia> y, spdMat, coeffϕ = DahlhausAR2(T, σ, μ)
julia> heatmap(log.(spdMat'), c = :viridis, clim = (-3,1.5), 
    size = (1000,400), title="True Spectrogram");
julia> plot(y, label = nothing, title = "Simulated time series from TvAR(2) model", 
        xlabel = "time", ylabel = "y")

TimeVaryingPeriodograms.simTvARMA — Function

simTvARMA(ϕ, θ, μ, σ)

Simulate a time-varying ARMA process

ϕ is a T×p matrix with time evolutions of the p AR coefficients

θ is a T×q matrix with time evolutions of the q MA coefficients

σ is a T×1 vector with time evolutions of the standard deviation of the noise

μ is a T×1 vector with time evolutions of the mean of the process

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> plot(y)