Index

TimeVaryingPeriodograms.BoundCorrectedMvPeriodogram

TimeVaryingPeriodograms.BoundCorrectedMvPeriodogram

TimeVaryingPeriodograms.DahlhausAR2

TimeVaryingPeriodograms.DahlhausAR2

TimeVaryingPeriodograms.EverittI

TimeVaryingPeriodograms.EverittI

TimeVaryingPeriodograms.SingleDft2

TimeVaryingPeriodograms.SingleDft2

TimeVaryingPeriodograms.SpecDensARMA

TimeVaryingPeriodograms.SpecDensARMA

TimeVaryingPeriodograms.completeDFTSingle

TimeVaryingPeriodograms.completeDFTSingle

TimeVaryingPeriodograms.completePeriodogram

TimeVaryingPeriodograms.completePeriodogram

TimeVaryingPeriodograms.mvPeriodogram

TimeVaryingPeriodograms.mvPeriodogram

TimeVaryingPeriodograms.mvPrewhitePeriodogram

TimeVaryingPeriodograms.mvPrewhitePeriodogram

TimeVaryingPeriodograms.mvPrewhiteTapPeriodogram

TimeVaryingPeriodograms.mvPrewhiteTapPeriodogram

TimeVaryingPeriodograms.mvTapPeriodogram

TimeVaryingPeriodograms.mvTapPeriodogram

TimeVaryingPeriodograms.prePeriodogram

TimeVaryingPeriodograms.prePeriodogram

TimeVaryingPeriodograms.preWhiteTvBlockPeriodogram

TimeVaryingPeriodograms.preWhiteTvBlockPeriodogram

TimeVaryingPeriodograms.predictiveDFT

TimeVaryingPeriodograms.predictiveDFT

TimeVaryingPeriodograms.reScaleTapTvBlockPeriodogram

TimeVaryingPeriodograms.reScaleTapTvBlockPeriodogram

TimeVaryingPeriodograms.simTvARMA

TimeVaryingPeriodograms.simTvARMA

TimeVaryingPeriodograms.tapBoundCorrectTvBlockPeriodogram

TimeVaryingPeriodograms.tapBoundCorrectTvBlockPeriodogram

TimeVaryingPeriodograms.tapBoundCorrectedMvPeriodogram

TimeVaryingPeriodograms.tapBoundCorrectedMvPeriodogram

TimeVaryingPeriodograms.tvPeriodogram

TimeVaryingPeriodograms.tvPeriodogram

TimeVaryingPeriodograms.whitePeriodogram

TimeVaryingPeriodograms.whitePeriodogram

TimeVaryingPeriodograms.wikleTV6AR

TimeVaryingPeriodograms.wikleTV6AR

Ramaining functions

TimeVaryingPeriodograms.BoundCorrectedMvPeriodogram — Function

    BoundCorrectedMvPeriodogram(y, m, p, startFreq) 

Same as mvPeriodogram but with boundary correction in Rao and Yang (2021). ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = BoundCorrectedMvPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.DahlhausAR2 — Function

DahlhausAR2(T, σ, μ)

Simulate spectrogram and time series from a time-varying autoregressive model of order 2 from Dahlhaus handbook chapter..

Length of time series is T, innovation standard deviation is σ. The mean of the time series is μ(default is 0).

Examples

julia> T = 1024; σ = 2; μ = 0;
julia> y, spdMat, coeffϕ = DahlhausAR2(T, σ, μ)
julia> heatmap(log.(spdMat'), c = :viridis, clim = (-3,1.5), 
    size = (1000,400), title="True Spectrogram");
julia> plot(y, label = nothing, title = "Simulated time series from TvAR(2) model", 
        xlabel = "time", ylabel = "y")

TimeVaryingPeriodograms.EverittI — Method

EverittI(y, ωgrid, ρ, ν)

Everitt et al. version of the pre-periodogram.

y - input time series,

ωgrid - grid of frequencies to be evaluated,

ρ - Parameter that controls smoothing ove time,

ν - parameter that controlls exponential taper.

TimeVaryingPeriodograms.SingleDft2 — Method

SingleDft2(y, fr)

Computing the DFT on a single frequency, fr. This is needed in the boundary correction for the dynamic Whittle likellihood..

Returns the discrete Fourier transform of the data.

TimeVaryingPeriodograms.SpecDensARMA — Method

SpecDensARMA(ω, ϕ, θ, σ²)

Compute spectral density for the univariate ARMA model.

ω is a radial frequency

ϕ is a vector of AR coefficients

θ is a vector of MA coefficients

σ² is the noise variance

Examples

The spectral density for an AR(1) process with unit noise variance is

julia> SpecDensARMA(0.5, 0.9, 0, 1)
0.6909224383713601

TimeVaryingPeriodograms.completeDFTSingle — Method

   completeDFTSingle(y, fr; p = nothing, ϕ = nothing)

Computing the complete DFT on a single frequency, fr. This is needed in the boundary correction for the dynamic Whittle likellihood.

Returns the discrete Fourier transform of the data.

TimeVaryingPeriodograms.completePeriodogram — Method

   completePeriodogram(y; fr=nothing, p = nothing, ϕ = nothing, α=nothing, threshold=0.1^8)

Computing the complete periodogram on frequencies, fr (Fourier frquencies if nothing). This is needed in the boundary correction for the dynamic Whittle likellihood.

Returns the complete periodogram of the data given the model.

TimeVaryingPeriodograms.mvPeriodogram — Method

  mvPeriodogram(y, m, startFreq) 

Computes the moving local periodogram of the time series y using a window size m, starting at frequency startFreq.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPeriodogram(y, 10, 2)

TimeVaryingPeriodograms.mvPrewhitePeriodogram — Function

  mvPrewhitePeriodogram(y, m, p=nothing, startFreq) 

Same as mvPeriodogram but with a Tukey taper applied to the data. Use α=1 for Hanning taper.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvTapPeriodogramDiffStartFreq(y, 10)

TimeVaryingPeriodograms.mvPrewhiteTapPeriodogram — Function

  mvPrewhiteTapPeriodogram(y, m, p, startFreq) 

Same as mvPrewhitePeriodogram but with prewhitening. p is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPreWPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.mvTapPeriodogram — Function

mvTapPeriodogram(y, m, α=1, startFreq=1) 

Same as mvPeriodogram but with a Tukey taper applied to the data. Use α=1 for Hanning taper.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvTapPeriodogramDiffStartFreq(y, 10)

TimeVaryingPeriodograms.prePeriodogram — Method

prePeriodogram(y, ω)

ω is a selected vector of frequencies. Returns a (length(ω) x T-1) matrix containing the prePeriodogram.

TimeVaryingPeriodograms.preWhiteTvBlockPeriodogram — Function

    preWhiteTvBlockPeriodogram(y,N,S,p=nothing)

Same as tvPeriodogram but with pre-whitening. ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = preWhiteTvBlockPeriodogram(y,25,10,p=nothing)

TimeVaryingPeriodograms.predictiveDFT — Method

  predictiveDFT(y, ω; p=nothing, ϕ=nothing)

Predictive DFT from Rao and Yang (2021), see r-package "cspec" on CRAN. Default is fitting the best AR model and compute the predictive DFT based on the predictions att frequencies ´ω´. ´p´ will fix the AR-order and ´ϕ´ gives the AR-coefficients.

Returns the discrete Fourier coefficients associated with the predictions.

Examples

julia> rand(Normal(500))
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = mvPreWPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.reScaleTapTvBlockPeriodogram — Function

    reScaleTapTvBlockPeriodogram(y,N,S)

Same as tvPeriodogram but with rescaling of the tapered data to match the variance of the non-scaled data. Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = reScaleTapTvBlockPeriodogram(y,25,10)

TimeVaryingPeriodograms.simTvARMA — Method

simTvARMA(ϕ, θ, μ, σ)

Simulate a time-varying ARMA process

ϕ is a T×p matrix with time evolutions of the p AR coefficients

θ is a T×q matrix with time evolutions of the q MA coefficients

σ is a T×1 vector with time evolutions of the standard deviation of the noise

μ is a T×1 vector with time evolutions of the mean of the process

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> plot(y)

TimeVaryingPeriodograms.tapBoundCorrectTvBlockPeriodogram — Function

    tapBoundCorrectTvBlockPeriodogram(y,N,S,p=nothing,α=nothing,thresH=0.1^8)

Same as tvPeriodogram but with boundary correction from Rao and Yang (2021). ´p´ is the lag length used in the AR-model, if left blanc the AR order is estimated on each segment.

Returns the local periodogram data MI_t at the Fourier frequencies.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI = tapBoundCorrectTvBlockPeriodogram(y,25,10,p=nothing,α=nothing,thresH=0.1^8)

TimeVaryingPeriodograms.tapBoundCorrectedMvPeriodogram — Function

tapBoundCorrectedMvPeriodogram(y, m, p, startFreq)

Same as ´BoundCorrectedMvPeriodogram´ but with tapering using a Hanning window.

Returns the local periodogram data MIt data and a vector λ with the frequencies where MIt is computed at frequency λ[1 + (t + startFreq - 2) mod m)].

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia> MI, λ = tapBoundCorrectedMvPeriodogram(y, 10, 2, 1)

TimeVaryingPeriodograms.tvPeriodogram — Function

    tvPeriodogram(y, N, S, nFreq = nothing, taper = nothing; taperArgs...)

Computes the block-periodogram (Dahlhaus, 2012) of the time series y using a window size N, and step-size S on the Fourier frequencies. NFreq can be used to evaluate other frequencies and taper can be used to apply a taper to the data, see package DSP for details.

Returns a spectrogram of the data.

Examples

julia> T = 500; ϕ = [sin.(2π*(1:T)/T) -0.1*ones(T)]; θ = zeros(T); σ = 1; μ = 2;
julia> y = simTvARMA(ϕ, θ, μ, σ);
julia>  tvPeriodogram(y, 25, 15)

TimeVaryingPeriodograms.whitePeriodogram — Method

whitePeriodogram(y, p)

Returns the periodogram from the prewhitening approach using AR-order p.

TimeVaryingPeriodograms.wikleTV6AR — Method

wikleTV6AR(T, σ, A, μ)

Simulate spectrogram and time series from a time-varying autoregressive model of order 6.

Length of time series is T, innovation standard deviation is σ, and the AR coefficients are compute from the three-element vector A. The mean of the time series is μ.

Examples

julia> T = 1024; σ = 2; A = [1.1, 1.12, 1.1]; μ = 0;
julia> y, spdMat, coeffϕ = wikleTV6AR(T, σ, A, μ)
julia> heatmap(log.(spdMat'), c = :viridis, clim = (-3,1.5), 
    size = (1000,400), title="True Spectrogram");
julia> plot(y, label = nothing, title = "Simulated time series from TvAR(6) model", 
        xlabel = "time", ylabel = "y")